+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 
 Changes in 0.9.X - DEV
 ======================
+- Set attribute to None does not work (at least for fields with default values) #734
 - Querying by a field defined in a subclass raises InvalidQueryError #744
 - Add Support For MongoDB 2.6.X's maxTimeMS #778
 - abstract shouldn't be inherited in EmbeddedDocument # 789